patch 8.2.4565: no command line completion for :breakadd and :breakdel
Problem:    No command line completion for :breakadd and :breakdel.
Solution:   Add completion for :breakadd and :breakdel. (Yegappan Lakshmanan,
            closes #9950)

+++ src/spellsuggest.c
@@ -508,7 +508,7 @@ spell_suggest(int count)
 	end_visual_mode();
 	// make sure we don't include the NUL at the end of the line
 	line = ml_get_curline();
-	if (badlen > STRLEN(line) - curwin->w_cursor.col)
+	if (badlen > (int)STRLEN(line) - (int)curwin->w_cursor.col)
 	    badlen = STRLEN(line) - curwin->w_cursor.col;
     }
     // Find the start of the badly spelled word.
